requirements.txt
fastapi
uvicorn
msal
httpx
pydantic
python-multipart
jinja2
python-dotenv
itsdangerous
